Пересматриват реализацию компонента комментариев
