+++ src/components/film-details.js
@@ -37,7 +37,7 @@ const createFilmDetailsRowTemplate = ([term, cell]) => {
 const createFilmDetailsTemplate = (film) => {
   const {poster, ageLimit, title, rating, director,
     writers, actors, release, duration, country,
-    description, genres, comments,
+    description, genres,
     isAddedToWatchlist, isMarkedAsWatched, isFavorite
   } = film;
 
@@ -88,13 +88,6 @@ const createFilmDetailsTemplate = (film) => {
      
            <input type="checkbox" class="film-details__control-input visually-hidden" id="favorite" name="favorite" ${makeInputActive(isFavorite)}>
            <label for="favorite" class="film-details__control-label film-details__control-label--favorite">Add to favorites</label>
-         </section>
-       </div>
-       <div class="form-details__bottom-container">
-         <section class="film-details__comments-wrap">
-           <h3 class="film-details__comments-title">Comments <span class="film-details__comments-count">${comments.length}</span></h3>
-           <ul class="film-details__comments-list">
-           </ul>
          </section>
        </div>
      </form>
